Centralize ts for outlet and details
